+++ src/clj_kondo/impl/namespace.clj
@@ -861,12 +861,38 @@
       (prn x)
       x)))
 
-;;;; Scratch
-
-(comment)
 
 (defn reg-protocol-impl! [{:keys [base-lang lang namespaces]} ns-name protocol-impl]
   (let [path [base-lang lang ns-name]]
     (swap! namespaces update-in path
            (fn [ns]
              (update ns :protocol-impls (fnil conj []) protocol-impl)))))
+
+(defn reg-defmulti! [{:keys [base-lang lang namespaces]} ns-sym multimethod-sym]
+  (let [path [base-lang lang ns-sym]]
+    (swap! namespaces update-in path
+           (fn [ns] (update ns :defmultis (fnil conj #{}) multimethod-sym)))))
+
+(defn reg-defmethod! [{:keys [lang] :as ctx} ns-sym multimethod-sym
+                      dispatch-val-node dispatch-val-str expr]
+  (let [key [multimethod-sym dispatch-val-str]
+        ns-ctx (get-in @(:namespaces ctx) [(:base-lang ctx) lang ns-sym])
+        multimethod-sym-ns-sym (some-> (namespace multimethod-sym) symbol)
+        qualified-resolved (and multimethod-sym-ns-sym
+                                (get-in ns-ctx [:qualify-ns multimethod-sym-ns-sym]))
+        resolved-defmulti-ns (or qualified-resolved
+                                 (and ((:defmultis ns-ctx #{}) multimethod-sym)
+                                      ns-sym))
+        metadata (assoc (meta expr)
+                        :filename (:filename ctx)
+                        :lang lang
+                        :dispatch-val-node dispatch-val-node
+                        :defmulti-ns resolved-defmulti-ns)]
+    (swap! (:namespaces ctx) update-in [(:base-lang ctx) lang ns-sym]
+           (fn [ns]
+             (update ns :defmethods (fnil conj [])
+                     [key metadata])))))
+
+(comment
+  ;;;; Scratch
+  )